+++ javascript/core/src/engine/verbs/chain.ts
@@ -3,7 +3,6 @@
  * Licensed under the MIT license. See LICENSE file in the project.
  */
 import { table } from 'arquero'
-import { uniqueId } from 'lodash-es'
 import type { StepFunction, TableStore } from '../../index.js'
 import type { ChainArgs, Step, TableContainer } from '../../types.js'
 import { aggregate } from './aggregate.js'
@@ -82,7 +81,7 @@ async function exec(step: Step, store: TableStore): Promise<TableContainer> {
 	const substore = nofork ? store : await store.clone()
 
 	let output: TableContainer = {
-		id: uniqueId(),
+		id: step.output,
 		name: step.output,
 		table: table({}),
 	}
@@ -104,6 +103,7 @@ async function exec(step: Step, store: TableStore): Promise<TableContainer> {
 	// parent store only gets the final output of the chain
 	const final = {
 		...output,
+		id: step.output,
 	}
 	store.set(final)
 	// return the final table
